scraper/requirements.txt
scrapy
chardet
psycopg2-binary
